refactor: Split logic from routes
